+++ src/main/java/com/dontforget/dontforget/app/notice/application/NoticeApplication.java
@@ -4,7 +4,6 @@ import com.dontforget.dontforget.app.notice.api.request.FCMNoticeRequest;
 import com.dontforget.dontforget.domain.notice.query.NoticeDeviceRequest;
 import com.dontforget.dontforget.domain.notice.service.AlarmSender;
 import com.dontforget.dontforget.domain.notice.service.CreateNoticeDevice;
-
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
