+++ src/main/java/com/dontforget/dontforget/config/FcmConfig.java
@@ -1,6 +1,5 @@
-package com.dontforget.dontforget.infra;
+package com.dontforget.dontforget.config;
 
-import com.dontforget.dontforget.domain.notice.service.AlarmExecutor;
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.FirebaseOptions;
@@ -8,16 +7,18 @@ import java.io.IOException;
 import java.io.InputStream;
 import javax.annotation.PostConstruct;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Profile;
 import org.springframework.core.io.ClassPathResource;
+import org.springframework.stereotype.Component;
 
 @Slf4j
-public class FCMExecutor implements AlarmExecutor {
+@Component
+@Profile("!test")
+public class FcmConfig {
 
-    private final String fcmKeyPath;
-
-    public FCMExecutor(String fcmKeyPath) {
-        this.fcmKeyPath = fcmKeyPath;
-    }
+    @Value("${firebase.key-path}")
+    private String fcmKeyPath;
 
     @PostConstruct
     public void initialize() {
@@ -36,5 +37,4 @@ public class FCMExecutor implements AlarmExecutor {
             throw new RuntimeException(e.getMessage());
         }
     }
-
 }
