--- src/main/java/com/dontforget/dontforget/domain/notice/service/AlarmExecutor.java
@@ -1,5 +1,0 @@
-package com.dontforget.dontforget.domain.notice.service;
-
-public interface AlarmExecutor {
-    void initialize();
-}
