--- src/test/java/com/dontforget/dontforget/config/FakeFcmExecutor.java
@@ -1,10 +1,0 @@
-package com.dontforget.dontforget.config;
-
-
-import com.dontforget.dontforget.domain.notice.service.AlarmExecutor;
-
-public class FakeFcmExecutor implements AlarmExecutor {
-    public void initialize() {
-        return;
-    }
-}
